split source into multiple files and refactor some more
